api/requirements.txt
fastapi[standard]==0.122.0
mariadb==1.1.14
python-dotenv==1.2.1
python-multipart
itsdangerous